Fix bug with DQ invalid datatype.

+++ romancal/outlier_detection/outlier_detection.py
@@ -384,8 +384,8 @@ def flag_cr(
     # Count existing DO_NOT_USE pixels
     count_existing = np.count_nonzero(sci_image.dq & DO_NOT_USE)
 
-    # Update the DQ array in the input image.
-    sci_image.dq = np.bitwise_or(sci_image.dq, cr_mask * (DO_NOT_USE | OUTLIER))
+    # Update the DQ array values in the input image but preserve datatype.
+    sci_image.dq = np.bitwise_or(sci_image.dq, cr_mask * (DO_NOT_USE | OUTLIER)).astype(np.uint32)
 
     # Report number (and percent) of new DO_NOT_USE pixels found
     count_outlier = np.count_nonzero(sci_image.dq & DO_NOT_USE)
